refactor 2. PrizeGrade의 패키지 domain으로 변경
