+++ src/main/java/lotto/README.md
@@ -11,3 +11,4 @@
 
 ## refactoring list
 1. 패키지명 변경 후 중복 소스인 step2 패키지 삭제
+2. PrizeGrade의 패키지 domain으로 변경
