+++ src/main/java/lotto/domain/ConfirmResults.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.constants.PrizeGrade;
-
 import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
