+++ src/main/java/lotto/domain/LottoGames.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.constants.PrizeGrade;
-
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
