+++ src/main/java/lotto/domain/PrizeGrade.java
@@ -1,4 +1,4 @@
-package lotto.constants;
+package lotto.domain;
 
 import java.util.Arrays;
