+++ src/test/java/lotto/constants/PrizeGradeTest.java
@@ -1,5 +1,6 @@
 package lotto.constants;
 
+import lotto.domain.PrizeGrade;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -8,7 +9,7 @@ import org.junit.jupiter.params.provider.MethodSource;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static lotto.constants.PrizeGrade.*;
+import static lotto.domain.PrizeGrade.*;
 
 class PrizeGradeTest {
