+++ src/test/java/lotto/domain/ConfirmResultsTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import lotto.constants.PrizeGrade;
 
 import java.util.Arrays;
 import java.util.List;
@@ -13,7 +12,7 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static lotto.constants.PrizeGrade.THIRD;
+import static lotto.domain.PrizeGrade.THIRD;
 
 class ConfirmResultsTest {
